+++ code/utils/xrLC/xrBuildRapidModel.cpp
@@ -49,7 +49,7 @@ void CBuild::BuildRapid		(BOOL bSaveForOtherCompilers)
 	xr_vector<Face*>			adjacent_vec;
 	adjacent_vec.reserve		(6*2*3);
 
-	CDB::CollectorPacked	CL	(scene_bb,lc_global_data()->g_vertices().size(),lc_global_data()->g_faces().size());
+	CDB::CollectorPacked	CL	(scene_bb, (u32)lc_global_data()->g_vertices().size(), (u32)lc_global_data()->g_faces().size());
 
 	for (vecFaceIt it=lc_global_data()->g_faces().begin(); it!=lc_global_data()->g_faces().end(); it++)
 	{
